Analysis Class

A class for defining Analysis details. Objects of class Analysis should not be created directly but by the constructor create_analysis_obj().

Slots

data_matrix

matrix. The data matrix, including all covariates to be adjusted for, all relevant outcome variables, and treatment arm and external control arm flags.

covariates

Covariate. Object of class Covariate as output by the function covariate_details().

outcome

Outcome. Object of class Outcome as output by outcome_surv_exponential(), outcome_surv_weibull_ph(), or outcome_bin_logistic().

borrowing

Borrowing. Object of class Borrowing as output by borrowing_full(), borrowing_none(), or borrowing_hierarchical_commensurate().

treatment

Treatment. Object of class Treatment as output by treatment_details().

model_string

character. The string that contains the full Stan model code to be compiled.

model

CmdStanModel. The compiled Stan model as output by cmdstanr::cmdstan_model()

ready_to_sample

logical. Is the object ready to sample?