PriorExponential Class

A class for defining exponential priors to be translated to Stan code. Objects of class PriorExponential should not be created directly but by the constructor prior_exponential().

Slots

stan_code

character. Stan implementation of the prior, with placeholders for exponential Stan function parameters surrounded with {{ and }} to be replaced with glue::glue().

n_param

integer. Number of prior parameters (1).

constraint

character. Support of prior distribution, "<lower=0>".

beta

numeric. Inverse scale (>0).

See also