Simulation Class

A class for defining Simulation study details. Objects of class Simulation should not be created directly but by the constructor create_simulation_obj().

Slots

data_matrix_list

SimDataList. The list of lists of data matrices created with sim_data_list().

outcome

SimOutcomeList. List of Outcome objects created with sim_outcome_list().

borrowing

SimBorrowingList. List of Borrowing objects created with sim_borrowing_list().

covariate

SimCovariateList or NULL. List of Covariate objects created with sim_covariate_list() or NULL (no covariate adjustment).

treatment

SimTreatmentList. List of Treatment objects created with sim_treatment_list().

guide

data.frame. Data.frame containing information on all combinations evaluated.

n_combos

integer. Number of combinations of parameters to be evaluated.

n_analyses

integer. Number of analyses (combos x datasets to be performed).

`analysis_obj_list`

list. List of analysis objects indexed according to guide.