UniformPrior Class

A class for defining uniform priors to be translated to Stan code. Objects of class UniformPrior should not be created directly but by the constructor uniform_prior().

Slots

stan_code

character. Stan implementation of the prior, with placeholders for uniform stan function parameters surrounded with {{ and }} to be replaced with glue::glue().

n_param

integer. Number of prior parameters (2).

constraint

character. Support of prior distribution, "<lower=alpha, upper=beta>".

alpha

numeric. Lower bound.

beta

numeric. Upper bound (>alpha).

See also