Full borrowing

Full borrowing

Usage

borrowing_full(ext_flag_col)

Arguments

ext_flag_col

character. Name of the external flag column in the matrix.

Value

Object of class BorrowingFull.

Details

Method

This method does not distinguish between internal and external arms, effectively pooling patients.

External Control

The ext_flag_col argument refers to the column in the data matrix that contains the flag indicating a patient is from the external control cohort.

See also

Examples

fb <- borrowing_full("ext")