Exponential survival distribution

Exponential survival distribution

Usage

outcome_surv_exponential(time_var, cens_var, baseline_prior, weight_var = "")

Arguments

time_var

character. Name of time variable column in model matrix

cens_var

character. Name of the censorship variable flag in model matrix

baseline_prior

Prior. Object of class Prior specifying prior distribution for the baseline outcome. See Details for more information.

weight_var

character. Optional name of variable in model matrix for weighting the log likelihood.

Value

Object of class OutcomeSurvExponential.

Details

Baseline Prior

The baseline_prior argument specifies the prior distribution for the baseline log hazard rate. The interpretation of the baseline_prior differs slightly between borrowing methods selected.

Dynamic borrowing using borrowing_hierarchical_commensurate(): the baseline_prior for Bayesian Dynamic Borrowing refers to the log hazard rate of the external control arm.

Full borrowing or No borrowing using borrowing_full() or borrowing_none(): the baseline_prior for these borrowing methods refers to the log hazard rate for the internal control arm.

See also

Examples

es <- outcome_surv_exponential(
  time_var = "time",
  cens_var = "cens",
  baseline_prior = prior_normal(0, 1000)
)