Input borrowing details for a simulation study

A function for defining which borrowing scenarios should be evaluated as part of a simulation study.

Usage

sim_borrowing_list(borrowing_list)

Arguments

borrowing_list

named list of objects of class Borrowing created by borrowing_full(), borrowing_none(), or borrowing_hierarchical_commensurate().

Value

Object of class SimBorrowingList.

See also

Examples

borrow_scenarios <- sim_borrowing_list(
  list(
    "No borrowing" = borrowing_none("ext"),
    "Full borrowing" = borrowing_full("ext"),
    "BDB, uninformative prior" = borrowing_hierarchical_commensurate(
      "ext",
      prior_gamma(0.001, 0.001)
    ),
    "BDB, informative prior" = borrowing_hierarchical_commensurate(
      "ext",
      prior_gamma(1, 0.001)
    )
  )
)