Input outcome details for a simulation study

A function for defining which outcome scenarios should be evaluated as part of a simulation study.

Usage

sim_outcome_list(outcome_list)

Arguments

outcome_list

named list of objects of class Outcome created by outcome_details().

Value

Object of class SimOutcomeList.

See also

Examples

outcome_scenarios <- sim_outcome_list(
  list(
    "Exponential" = outcome_surv_exponential("time", "cnsr", prior_normal(0, 10000))
  )
)